Crate bdk_coin_select

Structs

Bnb

Bnb represents the current state of the BnB algorithm.

BnbIter

CoinSelector

CoinSelector selects and deselects from a set of candidates.

CoinSelectorOpt

ExcessStrategy

Selection

SelectionError

WeightedValue

A WeightedValue represents an input candidate for CoinSelector. This can either be a single UTXO, or a group of UTXOs that should be spent together.

Enums

BnbLimit

Determines how we should limit rounds of branch and bound.

BranchStrategy

Strategy in which we should branch.

ExcessStrategyKind

SelectionConstraint

Constants

TXIN_BASE_WEIGHT

Txin “base” fields include outpoint (32+4) and nSequence (4). This does not include scriptSigLen or scriptSig.

Functions

coin_select_bnb

This is a variation of the Branch and Bound Coin Selection algorithm designed by Murch (as seen in Bitcoin Core).

Type Definitions

DecideStrategy

Closure to decide the branching strategy, alongside a score (if the current selection is a candidate solution).