Struct bdk_coin_select::Bnb

pub struct Bnb<'c, S> {
    pub pool: Vec<(usize, &'c WeightedValue)>,
    pub pool_pos: usize,
    pub best_score: S,
    pub selection: CoinSelector<'c>,
    pub rem_abs: u64,
    pub rem_eff: i64,
}

Bnb represents the current state of the BnB algorithm.

Fields
pool: Vec<(usize, &'c WeightedValue)>pool_pos: usizebest_score: Sselection: CoinSelector<'c>rem_abs: u64rem_eff: i64
Implementations

impl<'c, S: Ord> Bnb<'c, S>

pub fn new( selector: CoinSelector<'c>, pool: Vec<(usize, &'c WeightedValue)>, max: S ) -> Self

Creates a new Bnb.

pub fn into_iter<'f>( self, strategy: &'f DecideStrategy<'c, S> ) -> BnbIter<'c, 'f, S>

Turns our Bnb state into an iterator.

strategy should assess our current selection/node and determine the branching strategy and whether this selection is a candidate solution (if so, return the selection score).

pub fn backtrack(&mut self) -> bool

Attempt to backtrack to the previously selected node’s omission branch, return false otherwise (no more solutions).

pub fn forward(&mut self, skip: bool)

Continue down this branch and skip the inclusion branch if specified.

pub fn advertise_new_score(&mut self, score: S) -> bool

Compare the advertised score with the current best. The new best will be the smaller value. Return true if best is replaced.