Struct bdk_coin_select::CoinSelectorOpt

pub struct CoinSelectorOpt {
    pub target_value: Option<u64>,
    pub max_extra_target: u64,
    pub target_feerate: f32,
    pub long_term_feerate: Option<f32>,
    pub min_absolute_fee: u64,
    pub base_weight: u32,
    pub drain_weight: u32,
    pub spend_drain_weight: u32,
    pub min_drain_value: u64,
}

Fields
target_value: Option<u64>

The value we need to select. If the value is None, then the selection will be complete if it can pay for the drain output and satisfy the other constraints (e.g., minimum fees).

max_extra_target: u64

Additional leeway for the target value.

target_feerate: f32

The feerate we should try and achieve in sats per weight unit.

long_term_feerate: Option<f32>

The feerate

min_absolute_fee: u64

The minimum absolute fee. I.e., needed for RBF.

base_weight: u32

The weight of the template transaction, including fixed fields and outputs.

drain_weight: u32

Additional weight if we include the drain (change) output.

spend_drain_weight: u32

Weight of spending the drain (change) output in the future.

min_drain_value: u64

Minimum value allowed for a drain (change) output.

Implementations

impl CoinSelectorOpt

pub fn fund_outputs( txouts: &[TxOut], drain_output: &TxOut, drain_satisfaction_weight: u32 ) -> Self

pub fn long_term_feerate(&self) -> f32

pub fn drain_waste(&self) -> i64